+++ otterdog/operations/local_plan_operation.py
@@ -18,11 +18,9 @@ from .plan_operation import PlanOperation
 
 
 class LocalPlanOperation(PlanOperation):
-    def __init__(self, suffix: str, update_webhooks: bool, update_secrets: bool) -> None:
+    def __init__(self, suffix: str, update_webhooks: bool, update_secrets: bool, update_filter: str) -> None:
         super().__init__(
-            no_web_ui=False,
-            update_webhooks=update_webhooks,
-            update_secrets=update_secrets,
+            no_web_ui=False, update_webhooks=update_webhooks, update_secrets=update_secrets, update_filter=update_filter
         )
 
         self.suffix = suffix