+++ otterdog/operations/plan_operation.py
@@ -12,6 +12,7 @@ from colorama import Fore, Style
 
 from otterdog.config import OtterdogConfig
 from otterdog.models import ModelObject
+from otterdog.models.secret import Secret
 from otterdog.models.webhook import Webhook
 from otterdog.utils import IndentingPrinter, Change
 
@@ -19,8 +20,8 @@ from .diff_operation import DiffOperation, DiffStatus
 
 
 class PlanOperation(DiffOperation):
-    def __init__(self, no_web_ui: bool, update_webhooks: bool, update_secrets: bool):
-        super().__init__(no_web_ui, update_webhooks, update_secrets)
+    def __init__(self, no_web_ui: bool, update_webhooks: bool, update_secrets: bool, update_filter: str):
+        super().__init__(no_web_ui, update_webhooks, update_secrets, update_filter)
 
     def init(self, config: OtterdogConfig, printer: IndentingPrinter) -> None:
         super().init(config, printer)
@@ -78,8 +79,13 @@ class PlanOperation(DiffOperation):
         self.printer.println()
         model_header = current_object.get_model_header(parent_object)
 
-        # FIXME: this code should be moved to the Webhook model class.
-        redacted_keys = {"secret"} if self.resolve_secrets() is True else set()
+        # FIXME: this code should be moved to the model class.
+        if isinstance(expected_object, Secret):
+            redacted_keys = {"value"} if self.resolve_secrets() is True else set()
+        elif isinstance(expected_object, Webhook):
+            redacted_keys = {"secret"} if self.resolve_secrets() is True else set()
+        else:
+            redacted_keys = set()
 
         self.print_modified_dict(modified_object, model_header, redacted_keys, forced_update)